Added capability to read nested objects
Added capability for the engine to read nested objects inside the data
file. This can be used to organize data visually for the user and
hopefully will be used for more advanced functionality in the engine.

Also changed the data.js to data.json because I felt like it. Changed
index.html to match.