Added animations, edit README
Finally added the animation effects upon text generation and got them to work correctly. Currently implemented in CSS with a bit of jQuery to correctly sync them. An alternate implemenation using only jQuery is in the file as comments.

Added lines to index.html to have a backup of jQuery / Zepto if the CDN is down for either. Also index.html contains lines to use either jQuery / Zepto, with one set commented out.

Also added more entries to the database lists, and rewrote the README.